Moved dump_collection into utils.py to reuse it for operation dump #18 #2
